Move theme type in a separate file
