--- src/themes/fontSizes.ts
@@ -1,16 +1,0 @@
-export interface FontSizesHeading {
-  xxl: string;
-  xl: string;
-  lg: string;
-  md: string;
-  sm: string;
-  xs: string;
-}
-
-export interface FontSizesBody {
-  xl: string;
-  lg: string;
-  md: string;
-  sm: string;
-  xs: string;
-}
